#29: Renamed run_configuration package to just "runconfig". Changed minimum supported version of plugin to 2020.1. Fixed some runconfig text in AhkBundle.properties
