+++ build.gradle.kts
@@ -40,7 +40,7 @@ tasks.publishPlugin {
 
 // See https://github.com/JetBrains/gradle-intellij-plugin/
 intellij {
-    version = "2019.3"
+    version = "2020.1"
     type = "IC"
 }
