+++ src/main/java/de/nordgedanken/auto_hotkey/lang/psi/AhkTokenType.kt
@@ -1,8 +1,8 @@
 package de.nordgedanken.auto_hotkey.lang.psi
 
 import com.intellij.psi.tree.IElementType
-import com.intellij.psi.tree.TokenSet
 import de.nordgedanken.auto_hotkey.lang.core.AhkLanguage
+
 //import de.nordgedanken.auto_hotkey.psi.AHKTypes.*
 
 open class AhkTokenType(debugName: String) : IElementType(debugName, AhkLanguage)
@@ -13,7 +13,7 @@ open class AhkTokenType(debugName: String) : IElementType(debugName, AhkLanguage
 
 //val AHK_BLOCK_LIKE_EXPRESSIONS = tokenSetOf(BLOCK_EXPR)
 
-fun tokenSetOf(vararg tokens: IElementType) = TokenSet.create(*tokens)
+//fun tokenSetOf(vararg tokens: IElementType) = TokenSet.create(*tokens)
 
 //val AHK_COMMENTS = tokenSetOf(BLOCK_COMMENT, EOL_COMMENT)
