+++ src/main/java/de/nordgedanken/auto_hotkey/runconfig/ui/AhkSdkListCellRenderer.kt
@@ -1,4 +1,4 @@
-package de.nordgedanken.auto_hotkey.run_configurations.ui
+package de.nordgedanken.auto_hotkey.runconfig.ui
 
 import com.intellij.icons.AllIcons
 import com.intellij.openapi.projectRoots.Sdk
