+++ src/main/resources/localization/AhkBundle.properties
@@ -8,13 +8,13 @@ runconfig.configtab.scriptpath.filechooser.message=Please select the AutoHotkey 
 runconfig.configtab.scriptargs.label=Script &Arguments:
 runconfig.configtab.scriptrunner.label=Script &Runner:
 runconfig.configtab.scriptrunner.sdklabel.default=Project Default
-runconfig.configtab.scriptrunner.sdklabel.unrecognized=Unrecognized SDK
-runconfig.configtab.scriptrunner.sdklabel.noneselected=No AutoHotkey SDK found
+runconfig.configtab.scriptrunner.sdklabel.unrecognized=Unrecognized runner
+runconfig.configtab.scriptrunner.sdklabel.noneselected=No AutoHotkey runners found
 runconfig.configtab.scriptrunner.projectsettingsbutton.tooltip=Configure AutoHotkey Runners
 
 runconfig.configtab.error.scriptpath.notexist=The script path does not exist on disk!
 runconfig.configtab.error.scriptpath.notahkextension=The script path does not point to an AutoHotkey (.ahk) file!
-runconfig.configtab.error.scriptrunner.notahksdktype=The script runner does not point to a valid AutoHotkey SDK! (You can add a new SDK in the Project Structure dialog by clicking the folder icon next to the "Script Runner" dropdown and then clicking "New...")
+runconfig.configtab.error.scriptrunner.notahksdktype=The script runner does not point to a valid AutoHotkey runner! (Click the gear icon next to the "Script Runner" dropdown to open the AutoHotkey settings page where you can add a new runner)
 
 runconfig.general.info.label=Note: It is recommended to create/use a shortcut to rerun the config (eg Ctrl+Alt+R) during development for quick script reloading
